requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
click==8.2.0
Django==5.2.1
django-autoslug==1.9.9
django-cors-headers==4.7.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-nine==0.2.7
djangorestframework==3.16.0
elastic-transport==8.17.1
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
exceptiongroup==1.3.0
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uv==0.7.3
uvicorn==0.34.2
